fix: missing SvgIconProps type
