+++ packages/react/src/icons/InfoCircleIcon.tsx
@@ -1,4 +1,4 @@
-import type { SvgIconProps } from '../../../ui/types';
+import type { SvgIconProps } from '../types';
 
 export function InfoCircleIcon({
   theme,
