+++ packages/react/src/icons/NoFundIcon.tsx
@@ -1,4 +1,4 @@
-import type { SvgIconProps } from '../../types';
+import type { SvgIconProps } from '../types';
 
 export function NoFundIcon({ size, ...props }: SvgIconProps) {
   return (
