+++ packages/react/src/ui/Connect/components/ConnectorIcon.tsx
@@ -3,7 +3,7 @@ import type { ConnectorMetadata } from 'fuels';
 import { FuelWalletDevelopmentIcon } from '../../../icons/FuelWalletDevelopmentIcon';
 import { FuelWalletIcon } from '../../../icons/FuelWalletIcon';
 import { FueletIcon } from '../../../icons/FueletIcon';
-import type { SvgIconProps } from '../../types';
+import type { SvgIconProps } from '../../../types';
 import { getImageUrl } from '../utils/getImageUrl';
 
 type ConnectorIconProps = {
