--- packages/react/src/ui/types.ts
@@ -1,6 +1,0 @@
-export type SvgIconProps = {
-  theme?: string;
-  className?: string;
-  onClick?: () => void;
-  size: number;
-};
